start migration of inkcyclopedia to database
This is a somewhat tricky maneuver. All data up until now has been
dependent on the server context in which the data is being retrieved. As
such a dictionary has been a suitable way of saving this data. For the
ininkcyclopedia this is not the case, a list is much preferable to it
and fetching individual entries based on server, updating individual
entries etc is not very useful at all.

As such abc_table has been split up into abc_table_base and
abc_table_dict, and in addition to that an unfinished abc_table_list has
been added. Both of these classes inherit from abc_table_base with which
it has a number of methods in common.