created new upsert function in table base
This function has not been properly tested yet, but it should work and
it should be both more flexible and more easy to use than the
implementation used for the dictionary.

It still has some issues that needs to be ironed out when it comes to
reporting the results and handling errors (such as the new possibility
of an error where one or more of the required values are not specified).