remove degrees matching, improve number matching
The degrees matching is problematic because degrees are used for things
other than temperature and thus leaves a high probability of false
positives. Additionally the number matching part of the regex now
requires that the number does not start with a zero and that it is
preceded by a word boundary. As such it will no longer match '05c' or
'a5c', thus further reducing the risk of false positives.