remove cogbase
The cogbase was only used for mentions list which, this method was moved
to a default module in mrfreeze.lib meant to house various standardised
behaviours (such as this, a standardised way of printing a list of users
as mentions). As that was the only method from CogBase currently in use
that's the only method currently in default.

In addition, when testing that the cogbase could be removed I noticed
that !restart doesn't work because when restarting __main__.py can't
find MrFreeze. Probably a working directory issue, but since this cog
isn't really used anymore is seemed more reasonable to resolve this by
deleting it all together.