make banish TemplateEngine able to return template
This commit adds methods to the BanishTemplate and TemplateEngine
classes so that the template engine can determine which BanishTemplate
to use based on the invocation used when calling the command. If it's
also given a MuteResponseType it can decide which template to use and
return the appropriate template string.

As a proof of concept, the method for punishing unauthorized uses of
banish has been edited to rely on this template.