modularization of !banish: regular banishes work
Some code for modularizing the !banish command and migrating it to use
the new TemplateEngine. In its current state it can do banishes of
regular users (normal banish, micro, super, mega or with time
expressions), but it can not do custom responses for failed banishes
(such as trying to banish a mod or mr freeze) and it can not do
unbanishes.