add setproctitle, and use setproctitle
setproctitle added as dependency. This allows us to set the process
title when running the bot, which makes it easier to find the process in
htop or whatever process manager thing you're using.

Apparently it doesn't work as nice on Windows and you'll have to use
some 3rd party process explorer or whatever to get any benefit from it.
But that's really windows in a nutshell, innit?